1. Containerization/Le-Trong-Minh/backend/requirements.txt
fastapi == 0.65.1
uvicorn == 0.14.0
motor == 2.4.0
pytest == 7.3.1
httpx == 0.24.0
python-logstash